Interface Interactable, PlayerMove fixes, Lightpost script
Added IInteractable, an interface that requires a GameObject to tell it what it interacted with.
Added Lightpost script that utilizes IInteractable, and when interacted with will change its linked Light intensity to 1.
PlayerMove fixes, if the magnitude of the crossproduct of Cam.forward and 0,-1,0 (world down) goes less than 0.2F (aka the camera is mostly pointing down), switches the movement vector to camera.up (so, mostly pointing what you would expect "forward" to be).